coomments for changing sql : -making price range, hours, yelp rating and description optional because some are missing - added restaurant yelp id
