--- server/sqlHelpers.js
@@ -1,3 +1,0 @@
-var pg = require('pg');
-
-
